+++ tests/index.test.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import { render } from "@testing-library/react";
+
+import { App } from "~/app";
+
+describe("<App />", () => {
+  test("Should be able to mount component", async () => {
+    const { getByText } = render(<App />);
+
+    expect(getByText("Page Inspector").textContent).toBeTruthy();
+  });
+});