+++ webpack.config.js
@@ -7,11 +7,21 @@ module.exports = {
   entry: path.join(__dirname, "src", "index.tsx"),
   devtool: "inline-source-map",
   output: {
-    path: path.join(__dirname, "dist"),
-    filename: "bundle.js",
+    path: path.resolve(__dirname, "dist"),
+    filename: "[name].[contenthash].js",
+    chunkFilename: "[name].[contenthash].js",
+  },
+  optimization: {
+    runtimeChunk: "single",
+    splitChunks: {
+      chunks: "all",
+    },
   },
   devServer: {
     static: path.join(__dirname, "public"),
+    port: 3000,
+    hot: true,
+    historyApiFallback: true,
   },
   module: {
     rules: [
@@ -28,7 +38,10 @@ module.exports = {
     ],
   },
   resolve: {
-    extensions: [".tsx", ".ts", ".js"],
+    extensions: [".tsx", ".ts", ".js", "jsx"],
+    alias: {
+      "~": path.join(__dirname, "src"),
+    },
   },
   plugins: [
     new HtmlWebpackPlugin({